+++ packages/layout/src/components/SideNav/NavGroup/index.ts
@@ -1,0 +1,1 @@
+export { NavGroup as default } from './NavGroup';