+++ packages/layout/src/components/SideNav/SideNav.stories.tsx
@@ -21,25 +21,27 @@ export const Basic = () => (
         <Header />
         <SideNav.List>
             <SideNav.Nav path="/">
-                <DashboardIcon />
+                <DashboardIcon size="S" />
                 <Text>Dashboard</Text>
             </SideNav.Nav>
             <SideNav.Nav path="/reports">
-                <InsertChartIcon />
+                <InsertChartIcon size="S" />
                 <Text>Reports</Text>
             </SideNav.Nav>
             <SideNav.Nav path="/statements">
-                <FolderOpenIcon />
+                <FolderOpenIcon size="S" />
                 <Text>Statements</Text>
             </SideNav.Nav>
-            <SideNav.Nav path="/transactions">
-                <DateRangeIcon />
-                <Text>Transactions</Text>
-            </SideNav.Nav>
-            <SideNav.Nav path="/settings">
-                <SettingsIcon />
-                <Text>Settings</Text>
-            </SideNav.Nav>
+            <SideNav.Group title="Group">
+                <SideNav.Nav path="/transactions">
+                    <DateRangeIcon size="S" />
+                    <Text>Transactions</Text>
+                </SideNav.Nav>
+                <SideNav.Nav path="/settings">
+                    <SettingsIcon size="S" />
+                    <Text>Settings</Text>
+                </SideNav.Nav>
+            </SideNav.Group>
         </SideNav.List>
     </SideNav>
 );